backend/requirements.txt
openai
bcrypt
pymongo
jwt
flask
python-dotenv
flask-cors